move common/shared stub callback functions to helpers.h
lots of repetition.  probably still not efficient as all static functions, so lots of code duplication, but good first step
