+++ tests/test_log.c
@@ -31,50 +31,6 @@ static void __LOG_APPEND_ENTRIES_SEQ_ID(void *l, int count, int id, raft_term_t 
     }
 }
 
-static int __logentry_get_node_id(
-    raft_server_t* raft,
-    void *udata,
-    raft_entry_t *ety,
-    raft_index_t ety_idx
-    )
-{
-    return 0;
-}
-
-static int __log_offer(
-    raft_server_t* raft,
-    void *user_data,
-    raft_entry_t *entry,
-    raft_index_t entry_idx
-    )
-{
-    CuAssertIntEquals((CuTest*)raft, 1, entry_idx);
-    return 0;
-}
-
-static int __log_pop(
-    raft_server_t* raft,
-    void *user_data,
-    raft_entry_t *entry,
-    raft_index_t entry_idx
-    )
-{
-    raft_entry_t* copy = malloc(sizeof(*entry));
-    memcpy(copy, entry, sizeof(*entry));
-    llqueue_offer(user_data, copy);
-    return 0;
-}
-
-static int __log_pop_failing(
-    raft_server_t* raft,
-    void *user_data,
-    raft_entry_t *entry,
-    raft_index_t entry_idx
-    )
-{
-    return -1;
-}
-
 raft_cbs_t funcs = {
     .log_get_node_id = __logentry_get_node_id
 };
