+++ tests/test_server.c
@@ -16,87 +16,6 @@
 
 // TODO: leader doesn't timeout and cause election
 
-static int __raft_persist_term(
-    raft_server_t* raft,
-    void *udata,
-    raft_term_t term,
-    int vote
-    )
-{
-    return 0;
-}
-
-static int __raft_persist_vote(
-    raft_server_t* raft,
-    void *udata,
-    int vote
-    )
-{
-    return 0;
-}
-
-int __raft_applylog(
-    raft_server_t* raft,
-    void *udata,
-    raft_entry_t *ety,
-    raft_index_t idx
-    )
-{
-    return 0;
-}
-
-int __raft_applylog_shutdown(
-    raft_server_t* raft,
-    void *udata,
-    raft_entry_t *ety,
-    raft_index_t idx
-    )
-{
-    return RAFT_ERR_SHUTDOWN;
-}
-
-int __raft_send_requestvote(raft_server_t* raft,
-                            void* udata,
-                            raft_node_t* node,
-                            msg_requestvote_t* msg)
-{
-    return 0;
-}
-
-static int __raft_send_appendentries(raft_server_t* raft,
-                              void* udata,
-                              raft_node_t* node,
-                              msg_appendentries_t* msg)
-{
-    return 0;
-}
-
-static int __raft_log_get_node_id(raft_server_t* raft,
-        void *udata,
-        raft_entry_t *entry,
-        raft_index_t entry_idx)
-{
-    return atoi(entry->data);
-}
-
-static int __raft_log_offer(raft_server_t* raft,
-        void* udata,
-        raft_entry_t *entry,
-        raft_index_t entry_idx)
-{
-    return 0;
-}
-
-static int __raft_node_has_sufficient_logs(
-    raft_server_t* raft,
-    void *user_data,
-    raft_node_t* node)
-{
-    int *flag = (int*)user_data;
-    *flag += 1;
-    return 0;
-}
-
 raft_cbs_t generic_funcs = {
     .persist_term = __raft_persist_term,
     .persist_vote = __raft_persist_vote,
