+++ tests/test_snapshotting.c
@@ -13,71 +13,6 @@
 
 #include "helpers.h"
 
-static int __raft_persist_term(
-    raft_server_t* raft,
-    void *udata,
-    raft_term_t term,
-    int vote
-    )
-{
-    return 0;
-}
-
-static int __raft_persist_vote(
-    raft_server_t* raft,
-    void *udata,
-    int vote
-    )
-{
-    return 0;
-}
-
-static int __raft_applylog(
-    raft_server_t* raft,
-    void *udata,
-    raft_entry_t *ety,
-    raft_index_t idx
-    )
-{
-    return 0;
-}
-
-static int __raft_send_requestvote(raft_server_t* raft,
-                            void* udata,
-                            raft_node_t* node,
-                            msg_requestvote_t* msg)
-{
-    return 0;
-}
-
-static int __raft_send_appendentries(raft_server_t* raft,
-                              void* udata,
-                              raft_node_t* node,
-                              msg_appendentries_t* msg)
-{
-    return 0;
-}
-
-static int __raft_send_appendentries_capture(raft_server_t* raft,
-                              void* udata,
-                              raft_node_t* node,
-                              msg_appendentries_t* msg)
-{
-    msg_appendentries_t* msg_captured = (msg_appendentries_t*)udata;
-    memcpy(msg_captured, msg, sizeof(msg_appendentries_t));
-    return 0;
-}
-
-static int __raft_send_snapshot_increment(raft_server_t* raft,
-        void* udata,
-        raft_node_t* node)
-{
-    int *counter = udata;
-
-    (*counter)++;
-    return 0;
-}
-
 /* static raft_cbs_t generic_funcs = { */
 /*     .persist_term = __raft_persist_term, */
 /*     .persist_vote = __raft_persist_vote, */
